+++ src/components/desktop/PassengersFormInfo.vue
@@ -42,8 +42,8 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "src/assets/variables.scss";
-@import "src/assets/font.scss";
+@import "@/assets/variables.scss";
+@import "@/assets/font.scss";
 
 .passengers-form-info-section {
   margin-left: 88px;
