+++ src/views/Ticket-booking.vue
@@ -26,9 +26,9 @@ import Footer from "@/components/Footer";
 import LoginModal from "@/components/LoginModal";
 import RegistrationModal from "@/components/RegistrationModal";
 import ForgotPasswordModal from "@/components/ForgotPasswordModal";
-import ThereBackTabs from "@/components/ThereBackTabs";
+import ThereBackTabs from "@/components/desktop/ThereBackTabs";
 import CurrentPoints from "@/components/CurrentPoints";
-import PassengersFormInfo from "@/components/PassengersFormInfo";
+import PassengersFormInfo from "@/components/desktop/PassengersFormInfo";
 import RemoveModal from "@/components/RemoveModal";
 import TimelineMobile from "@/components/mobile/TimelineMobile";
 import ThereTabMobile from "@/components/mobile/ThereTabMobile";
